+++ apps/server/prisma/migrations/20260228044619_soft_delete_machine/migration.sql
@@ -1,0 +1,2 @@
+-- AlterTable
+ALTER TABLE "Machine" ADD COLUMN     "deletedAt" TIMESTAMP(3);